+++ RealmServer/RealmServerClass.cs
@@ -9,6 +9,7 @@ namespace RealmServer
 {
     public class RealmServerClass : IDisposable
     {
+        public static List<RealmServerSession> Sessions = new List<RealmServerSession>();
         private readonly Socket _socketHandler;
 
         public RealmServerClass(IPEndPoint authPoint)
@@ -17,7 +18,7 @@ namespace RealmServer
             _socketHandler = new Socket(AddressFamily.InterNetwork, SocketType.Stream, ProtocolType.Tcp);
             _socketHandler.SetSocketOption(SocketOptionLevel.Tcp, SocketOptionName.NoDelay, true);
             _socketHandler.SetSocketOption(SocketOptionLevel.Socket, SocketOptionName.DontLinger, true);
-            _socketHandler.SetSocketOption(SocketOptionLevel.Socket, SocketOptionName.KeepAlive, false);
+            //_socketHandler.SetSocketOption(SocketOptionLevel.Socket, SocketOptionName.KeepAlive, true);
 
             _socketHandler.SendTimeout = 3500;
             _socketHandler.ReceiveTimeout = 3500;
@@ -57,7 +58,9 @@ namespace RealmServer
             var connectionSocket = ((Socket) asyncResult.AsyncState).EndAccept(asyncResult);
             var connectionId = GetFreeId();
 
-            ActiveConnections.Add(connectionId, new RealmServerSession(connectionId, connectionSocket));
+            var session = new RealmServerSession(connectionId, connectionSocket);
+            RealmServerSession.Sessions.Add(session);
+            ActiveConnections.Add(connectionId, session);
             _socketHandler.BeginAccept(ConnectionRequest, _socketHandler);
         }
 
@@ -67,7 +70,7 @@ namespace RealmServer
         /// <returns>int</returns>
         private int GetFreeId()
         {
-            for (var i = 0; i < 150; i++)
+            for (var i = 0; i < 3500; i++)
                 if (!ActiveConnections.ContainsKey(i))
                     return i;