+++ RealmServer/RealmServerDatabase.cs
@@ -1,0 +1,21 @@
+﻿using System.Linq;
+using Common.Database;
+using Common.Database.Tables;
+
+namespace RealmServer
+{
+    public class RealmServerDatabase : DatabaseModel<Models>
+    {
+        /// <summary>
+        ///     Get account based on login
+        /// </summary>
+        /// <param name="username"></param>
+        /// <returns></returns>
+        internal Users GetAccount(string username)
+        {
+            return !Model.Users.Any()
+                ? null
+                : Model.Users.FirstOrDefault(a => a.username.ToLower() == username.ToLower());
+        }
+    }
+}