+++ RealmServer/RealmServerRouter.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using Common.Globals;
 using Common.Helpers;
 
@@ -7,6 +8,8 @@ namespace RealmServer
     public class RealmServerRouter
     {
         public delegate void ProcessRealmPacketCallback(RealmServerSession session, byte[] data);
+
+        public delegate void ProcessRealmPacketCallback<in T>(RealmServerSession session, T handler);
 
         public static readonly Dictionary<RealmEnums, ProcessRealmPacketCallback> MCallbacks =
             new Dictionary<RealmEnums, ProcessRealmPacketCallback>();
@@ -26,8 +29,22 @@ namespace RealmServer
             else
             {
                 Log.Print(LogType.RealmServer, $"Missing handler: {opcode}");
-                RealmServerSession.DumpPacket(data, realmServerSession);
+                Utils.DumpPacket(data);
             }
+        }
+
+        internal static void AddHandler(RealmEnums opcode, ProcessRealmPacketCallback handler)
+        {
+            MCallbacks.Add(opcode, handler);
+        }
+
+        internal static void AddHandler<T>(RealmEnums opcode, ProcessRealmPacketCallback<T> callback)
+        {
+            AddHandler(opcode, (session, data) =>
+            {
+                var generatedHandler = (T) Activator.CreateInstance(typeof(T), data);
+                callback(session, generatedHandler);
+            });
         }
     }
 }